\name{authoryear_t}
\alias{authoryear_t}
\title{format the author and year}
\usage{
  authoryear_t(entry,
    format_inline_fn = format_authoryear_t)
}
\arguments{
  \item{entry}{a bibentry}

  \item{format_inline_fn}{the function that actually
  creates the inline format for a single entry}
}
\value{
  the author-year citation
}
\description{
  format the author and year
}

