% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cite_options}
\alias{cite_options}
\title{Toggles for various citation options}
\usage{
cite_options(citation_format = "compatibility", ...)
}
\arguments{
\item{citation_format}{'pandoc', 'compatibility' (with version 0.5 or earlier), or 'text'}

\item{...}{additional arguments to BibOptions}
}
\value{
updates the option specified for the duration of the session.
}
\description{
Toggles for various citation options
}

