\name{greycite}
\alias{greycite}
\title{greycite citation method}
\usage{
greycite(url, format = c("bib", "json"))
}
\arguments{
  \item{url}{the website URL}

  \item{format}{have greycite return data in bib (bibtex)
  or json format?}
}
\value{
  a bibentry class object
}
\description{
  Grab citation data using the greycite API
}

