% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numeral}
\alias{numeral}
\title{Formatting for numeric inline citations.  Needs testing.}
\usage{
numeral(entry)
}
\arguments{
\item{entry}{a bibentry object, e.g. from cite}
}
\description{
Formatting for numeric inline citations.  Needs testing.
}

