% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.bibtex}
\alias{read.bibtex}
\title{read.bibtex}
\usage{
read.bibtex(file, .Encoding = "UTF-8", header = if (length(preamble))
  paste(preamble, sep = "\\n") else "", footer = "",
  check = BibOptions()$check.entries)
}
\description{
read.bibtex
}

