% Generated by roxygen2 (4.0.1): do not edit by hand
\name{record_as_cited}
\alias{record_as_cited}
\title{record as cited}
\usage{
record_as_cited(entry)
}
\arguments{
\item{entry}{a BibEntry object (or something that can be coerced to one, like bibentry)}
}
\value{
Adds the object to the citation list, with a BibTex citation
key that does not duplicate any that are already in the record.
Invisibly returns the bibentry object.
}
\description{
adds a BibEntry object to the works cited list even if not cited
with a call to citep/citet
}

