% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bibliography}
\alias{bibliography}
\title{Generate the bibliography}
\usage{
bibliography(..., style = NULL)
}
\arguments{
\item{...}{additional arguments passed to PrintBibliography,}
}
\value{
a list of works cited
}
\description{
Generate the bibliography
}

