\name{cite}
\alias{cite}
\title{Add a citation (internal for citet and citep)}
\usage{
  cite(x, inline_format)
}
\arguments{
  \item{x}{a doi or list of dois, or a bibentry (or list of
  bibentries)}
}
\value{
  a parenthetical inline citation
}
\description{
  Add a citation (internal for citet and citep)
}
\details{
  Stores the full citation in a "works_cited" list, which
  can be printed with \code{\link{bibliography}}
}
\keyword{internal}

