% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanbib.R
\name{cleanbib}
\alias{Clean}
\alias{cited}
\alias{cleanbib}
\alias{far.}
\alias{log}
\alias{of}
\alias{reset_bib}
\alias{so}
\alias{the}
\alias{works}
\title{cleanbib}
\usage{
cleanbib()
}
\description{
cleanbib
}

