\name{formatref}
\alias{formatref}
\title{Convert crossref XML into a bibentry object}
\usage{
  formatref(a)
}
\arguments{
  \item{a}{crossref XML output}
}
\value{
  a bibentry format of the output
}
\description{
  Convert crossref XML into a bibentry object
}
\details{
  internal helper function
}
\keyword{internal}

