% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/greycite.R
\name{greycite}
\alias{greycite}
\title{greycite citation method}
\usage{
greycite(url, format = c("bib", "json"))
}
\arguments{
\item{url}{the website URL}

\item{format}{have greycite return data in bib (bibtex) or json format?}
}
\value{
a bibentry class object
}
\description{
Grab citation data using the greycite API
}

