% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.bibtex.R
\name{read.bibtex}
\alias{read.bibtex}
\title{read.bibtex}
\usage{
read.bibtex(file, .Encoding = "UTF-8", check = FALSE)
}
\arguments{
\item{file}{string; bib file to parse.}

\item{.Encoding}{encoding}

\item{check}{error, warn, or logical FALSE.  What action should be
taken if an entry is missing required fields?  FALSE means
no checking is done, warn means entry is added with an
error. error means the entry will not be added.  See
BibOptions.}
}
\description{
read.bibtex
}
